Removido velho pacote de persistência
