+++ src/main/kotlin/authentication/persistence/dao/UserDAO.kt
@@ -1,11 +1,10 @@
 package authentication.persistence.dao
 
-import authentication.Permission
 import authentication.User
 import authentication.persistence.tables.PermissionTable
 import authentication.persistence.tables.UserTable
-import newPersistence.DAO
-import newPersistence.ResultSet
+import persistence.DAO
+import persistence.ResultSet
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.transaction
 import java.util.*
