+++ src/main/kotlin/blockchain/persistence/dao/TransactionDAO.kt
@@ -1,14 +1,10 @@
 package blockchain.persistence.dao
 
-import authentication.persistence.tables.UserSessionTable
 import blockchain.Transaction
-import blockchain.persistence.tables.NodeInfoTable
 import blockchain.persistence.tables.TransactionTable
-import newPersistence.DAO
-import newPersistence.ResultSet
-import org.jetbrains.exposed.dao.id.EntityID
+import persistence.DAO
+import persistence.ResultSet
 import org.jetbrains.exposed.sql.*
-import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.transactions.transaction
 import java.util.*
 import kotlin.math.ceil
