+++ src/main/kotlin/document/persistence/dao/person/OfficialDAO.kt
@@ -3,8 +3,8 @@ package document.persistence.dao.person
 import document.persistence.tables.person.OfficialTable
 import document.persistence.tables.person.PersonTable
 import document.person.Official
-import newPersistence.DAO
-import newPersistence.ResultSet
+import persistence.DAO
+import persistence.ResultSet
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.transaction
 import java.util.*
