+++ src/main/kotlin/document/persistence/dao/person/PhysicalPersonDAO.kt
@@ -3,8 +3,8 @@ package document.persistence.dao.person
 import document.persistence.tables.person.PersonTable
 import document.persistence.tables.person.PhysicalPersonTable
 import document.person.PhysicalPerson
-import newPersistence.DAO
-import newPersistence.ResultSet
+import persistence.DAO
+import persistence.ResultSet
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.transaction
 import java.util.*
