+++ src/main/kotlin/persistence/DAO.kt
@@ -1,5 +1,17 @@
 package persistence
 
-interface DAO<Type> {
-    fun toType():Type?
+import org.jetbrains.exposed.sql.Op
+import org.jetbrains.exposed.sql.ResultRow
+
+interface DAO<Type, ID:Comparable<ID>>{
+    fun insert(obj: Type): Type
+    fun select(id: ID): Type?
+    fun selectMany(condition: Op<Boolean>, page:Int=1, pageLength:Int=20): ResultSet<Type>
+    fun selectMany(condition: Op<Boolean>): List<Type>
+    fun selectAll(page:Int=1, pageLength:Int=20): ResultSet<Type>
+    fun update(obj: Type)
+    fun remove(id: ID)
+    fun removeWhere(condition: Op<Boolean>)
+
+    fun toType(row: ResultRow):Type
 }
