+++ src/main/kotlin/persistence/ResultSet.kt
@@ -1,4 +1,4 @@
-package newPersistence
+package persistence
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
